requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==5.0.6
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
environs==11.0.0
flake8==7.0.0
flake8-broken-line==1.0.0
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
gunicorn==22.0.0
idna==3.7
kombu==5.3.7
marshmallow==3.21.3
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.1
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
redis==5.0.6
requests==2.32.3
requests-oauthlib==2.0.0
secure-smtplib==0.1.1
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
